Reverted separated libs change. Fixed directories tree. Cleaned CMakeLists.txt.
