+++ tests/common/serenity.hpp
@@ -9,6 +9,7 @@
 #include <mesos/mesos.pb.h>
 
 namespace mesos {
+namespace serenity {
 namespace tests {
 
 
@@ -87,7 +88,8 @@ private:
   ResourceUsage results;
 };
 
-} // tests {
-} // mesos {
+} // namespace tests {
+} // namespace serenity {
+} // namespace mesos {
 
 #endif //TESTS_SERENITY_HPP
