+++ tests/common/sinks/dummy_sink.hpp
@@ -7,13 +7,14 @@
 
 namespace mesos {
 namespace serenity {
+namespace tests {
 
 template<typename T>
 class DummySink : public Consumer<T>
 {
 public:
 
-  DummySink<T>() : numberOfMessagesConsumed(0){};
+  DummySink<T>() : numberOfMessagesConsumed(0) {};
 
   Try<Nothing> consume(T& in)
   {
@@ -24,7 +25,8 @@ public:
   int numberOfMessagesConsumed;
 };
 
-} // namespace serenity
-} // namespace mesos
+} // namespace tests {
+} // namespace serenity {
+} // namespace mesos {
 
 #endif //SERENITY_DUMMY_RESOURCE_USAGE_SINK_HPP
