+++ tests/common/sources/json_source.proto
