+++ tests/controllers/qos_controller_test.cpp
@@ -13,7 +13,7 @@
 
 #include "qos_controller/serenity_controller.hpp"
 
-#include "tests/serenity.hpp"
+#include "tests/common/serenity.hpp"
 
 using std::list;
 
@@ -21,6 +21,7 @@ using mesos::slave::QoSController;
 using mesos::slave::QoSCorrection;
 
 namespace mesos {
+namespace serenity {
 namespace tests {
 
 // NOTE: For now checking only the interface.
@@ -44,5 +45,7 @@ TEST(SerenityControllerTest, NoQoSCorrections)
   EXPECT_EQ(0u, result.get().size());
 }
 
-} // tests {
-} // mesos {
+} // namespace tests {
+} // namespace serenity {
+} // namespace mesos {
+
