+++ tests/filters/moving_average_test.cpp
